Simplify models down to one file
