--- project/models/trials.py
@@ -1,72 +1,0 @@
-from datetime import date
-from enum import auto
-
-from pydantic import BaseModel, Extra, Field
-
-from boilerdata.enums import GetNameEnum
-
-
-class Rod(GetNameEnum):
-    """The rod used in this trial."""
-
-    W = auto()
-    X = auto()
-    Y = auto()
-
-
-class Coupon(GetNameEnum):
-    """The coupon attached to the rod for this trial."""
-
-    A1 = auto()
-    A2 = auto()
-    A3 = auto()
-    A4 = auto()
-    A6 = auto()
-    A7 = auto()
-    A9 = auto()
-
-
-class Sample(GetNameEnum):
-    """The sample attached to the coupon in this trial."""
-
-    NA = auto()  # If no sample is attached to the coupon.
-    B3 = auto()
-
-
-class Group(GetNameEnum):
-    """The group that this sample belongs to."""
-
-    control = auto()
-    porous = auto()
-    hybrid = auto()
-
-
-class Joint(GetNameEnum):
-    """The method used to join parts of the sample in this trial."""
-
-    paste = auto()
-    epoxy = auto()
-    solder = auto()
-
-
-# * -------------------------------------------------------------------------------- * #
-# * MODELS
-
-
-class Trial(BaseModel, extra=Extra.forbid):
-    """A trial."""
-
-    date: date
-    rod: Rod
-    coupon: Coupon
-    sample: Sample
-    group: Group
-    monotonic: bool = Field(..., description="Whether the boiling curve is monotonic.")
-    joint: Joint
-    comment: str
-
-
-class Trials(BaseModel):
-    """The trials."""
-
-    trials: list[Trial]
